Update KTG029 add request list 198

+++ nts.uk/uk.at/at.screen/nts.uk.screen.at.app/src/main/java/nts/uk/screen/at/app/ktgwidget/find/dto/YearlyHoliday.java
@@ -22,16 +22,16 @@ public class YearlyHoliday {
 	
 	/*次回付与日*/
 	private GeneralDate nextGrantDate = GeneralDate.today();
-	
+	/*付与前*/
 	private YearlyHolidayInfo nextGrantDateInfo  = new YearlyHolidayInfo();
-	
+	/*付与後*/
 	private YearlyHolidayInfo afterGrantDateInfo = new YearlyHolidayInfo();
 	
 	/*出勤率*/
-	private int attendanceRate = 0;
+	private Double attendanceRate = 0.0;
 	
 	/*労働日数*/
-	private int workingDays = 0;
+	private Double workingDays = 0.0;
 	
 	/* 計算方法 */
 	private int calculationMethod = 0;
